Fixing bug in color output

+++ src/theia/io/write_bundler_files.cc
@@ -116,8 +116,9 @@ bool WriteBundleFile(const Reconstruction& reconstruction,
     const Eigen::Vector3d position = track->Point().hnormalized();
     ofs_bundle << position.transpose().format(unaligned) << std::endl;
 
-    ofs_bundle << track->Color()[0] << " " << track->Color()[1] << " "
-               << track->Color()[2] << std::endl;
+    ofs_bundle << track->Color().cast<double>()[0] << " "
+               << track->Color().cast<double>()[1] << " "
+               << track->Color().cast<double>()[2] << std::endl;
     const auto& views_in_track = track->ViewIds();
     ofs_bundle << views_in_track.size();
     for (const ViewId view_id : views_in_track) {
